Created a Formula for ICU4C.
There was a weird bit in getting the ICU4C configure script to recognize
the endianess of the machine. Apparently it was just checking if -arch
was provided as a build option and using universal endianess which would
cause the build to fail.

I haven't tested this on a PPC Mac but assuming ICU's scripts work this
should too.